// packages/defi_orders_frontend/codegen/UniversalSwap.ts
export type AssetStruct = {
  pool: PromiseOrValue<string>;
  manager: PromiseOrValue<string>;
  tokenId: PromiseOrValue<BigNumberish>;
  liquidity: PromiseOrValue<BigNumberish>;
  data: PromiseOrValue<BytesLike>;
};
// ...
export type DesiredStruct = {
  outputERC20s: PromiseOrValue<string>[];
  outputERC721s: AssetStruct[];
  ratios: PromiseOrValue<BigNumberish>[];
  minAmountsOut: PromiseOrValue<BigNumberish>[];
};